requirements.txt
asgiref==3.7.2
attrs==22.1.0
CacheControl==0.12.11
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==2.0.1
crashtest==0.4.1
cryptography==37.0.4
distlib==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
dulwich==0.20.46
filelock==3.8.0
html5lib==1.1
idna==3.4
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
more-itertools==8.14.0
msgpack==1.0.4
packaging==22.0
pexpect==4.8.0
pipenv==2022.12.19
pkginfo==1.9.2
platformdirs==2.5.2
poetry==1.3.2
poetry-core==1.4.0
poetry-plugin-export==1.2.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
pyrsistent==0.18.1
python-dotenv==1.0.0
pytz==2023.3
rapidfuzz==2.13.7
requests==2.28.1
requests-toolbelt==0.9.1
SecretStorage==3.3.3
shellingham==1.5.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.5
trove-classifiers==2022.12.22
typing_extensions==4.7.1
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
